services/lobby/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0
kafka-python-ng==2.2.3
psycopg[binary]==3.2.9


